fix route, create single typeahead component
