+++ console/console-init/ui/src/components/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./TablePagination";
+export * from "./TypeAheadSelect";
