+++ console/console-init/ui/src/modules/address-detail/containers/AddressLinks/AddressLinksPage.tsx
@@ -8,17 +8,17 @@ import {
   PageSection,
   PageSectionVariants,
   Title,
-  Pagination,
   Grid,
   GridItem
 } from "@patternfly/react-core";
 import { GridStylesForTableHeader } from "modules/address/AddressPage";
-import { AddressLinksContainer } from "../AddressLinksContainer";
-import { useHistory, useLocation } from "react-router";
-import { css } from "emotion";
-import { AddressLinksToolbar } from "modules/address-detail/containers/AddressLinksToolbar/AddressLinksToolbar";
+import { AddressLinksContainer } from "modules/address-detail/containers/AddressLinksContainer";
+import { useLocation } from "react-router";
+import { AddressLinksToolbar } from "modules/address-detail/containers/AddressLinksToolbar";
 import { ISortBy } from "@patternfly/react-table";
 import { Divider } from "@patternfly/react-core/dist/js/experimental";
+import { css } from "@patternfly/react-styles";
+import { TablePagination } from "components/TablePagination";
 
 interface IAddressLinksListPageProps {
   addressspace_name: string;
@@ -36,59 +36,19 @@ const AddressLinksPage: React.FunctionComponent<IAddressLinksListPageProps> = ({
   addressDisplayName
 }) => {
   const location = useLocation();
-  const history = useHistory();
   const searchParams = new URLSearchParams(location.search);
   const page = parseInt(searchParams.get("page") || "", 10) || 1;
   const perPage = parseInt(searchParams.get("perPage") || "", 10) || 10;
   const [addresLinksTotal, setAddressLinksTotal] = React.useState<number>(0);
   const [filterValue, setFilterValue] = React.useState<string>("Name");
-  const [filterNames, setFilterNames] = React.useState<Array<any>>([]);
+  const [filterNames, setFilterNames] = React.useState<
+    Array<{ value: string; isExact: boolean }>
+  >([]);
   const [sortDropDownValue, setSortDropdownValue] = React.useState<ISortBy>();
-  const [filterContainers, setFilterContainers] = React.useState<Array<any>>(
-    []
-  );
+  const [filterContainers, setFilterContainers] = React.useState<
+    Array<{ value: string; isExact: boolean }>
+  >([]);
   const [filterRole, setFilterRole] = React.useState<string>();
-
-  const setSearchParam = React.useCallback(
-    (name: string, value: string) => {
-      searchParams.set(name, value.toString());
-    },
-    [searchParams]
-  );
-
-  const handlePageChange = React.useCallback(
-    (_: any, newPage: number) => {
-      setSearchParam("page", newPage.toString());
-      history.push({
-        search: searchParams.toString()
-      });
-    },
-    [setSearchParam, history, searchParams]
-  );
-
-  const handlePerPageChange = React.useCallback(
-    (_: any, newPerPage: number) => {
-      setSearchParam("page", "1");
-      setSearchParam("perPage", newPerPage.toString());
-      history.push({
-        search: searchParams.toString()
-      });
-    },
-    [setSearchParam, history, searchParams]
-  );
-
-  const renderPagination = (page: number, perPage: number) => {
-    return (
-      <Pagination
-        itemCount={addresLinksTotal}
-        perPage={perPage}
-        page={page}
-        onSetPage={handlePageChange}
-        variant="top"
-        onPerPageSelect={handlePerPageChange}
-      />
-    );
-  };
 
   return (
     <PageSection>
@@ -118,7 +78,12 @@ const AddressLinksPage: React.FunctionComponent<IAddressLinksListPageProps> = ({
             />
           </GridItem>
           <GridItem span={5}>
-            {addresLinksTotal > 0 && renderPagination(page, perPage)}
+            <TablePagination
+              itemCount={addresLinksTotal}
+              variant={"top"}
+              page={page}
+              perPage={perPage}
+            />
           </GridItem>
         </Grid>
         <Divider />
@@ -136,7 +101,12 @@ const AddressLinksPage: React.FunctionComponent<IAddressLinksListPageProps> = ({
           setSortValue={setSortDropdownValue}
           filterRole={filterRole}
         />
-        {addresLinksTotal > 0 && renderPagination(page, perPage)}
+        <TablePagination
+          itemCount={addresLinksTotal}
+          variant={"bottom"}
+          page={page}
+          perPage={perPage}
+        />
       </PageSection>
     </PageSection>
   );
