+++ console/console-init/ui/src/modules/address-detail/containers/AddressLinksToolbar/AddressLinksToolbar.tsx
@@ -5,48 +5,15 @@
 
 import * as React from "react";
 import {
-  DataToolbarGroup,
-  DataToolbarFilter,
-  DataToolbarItem,
-  DataToolbarChip,
-  DataToolbarToggleGroup,
   DataToolbar,
   DataToolbarContent
 } from "@patternfly/react-core/dist/js/experimental";
-import {
-  Dropdown,
-  DropdownToggle,
-  DropdownItem,
-  InputGroup,
-  Button,
-  ButtonVariant,
-  Badge,
-  Select,
-  SelectVariant,
-  SelectOption,
-  SelectOptionObject
-} from "@patternfly/react-core";
-import { FilterIcon, SearchIcon } from "@patternfly/react-icons";
 import { ISortBy } from "@patternfly/react-table";
 import useWindowDimensions from "components/common/WindowDimension";
 import { SortForMobileView } from "components/common/SortForMobileView";
-import { useApolloClient } from "@apollo/react-hooks";
-import {
-  RETURN_ALL_CONTAINER_IDS_OF_ADDRESS_LINKS_FOR_TYPEAHEAD_SEARCH,
-  RETURN_ALL_NAMES_OF_ADDRESS_LINK_FOR_TYPEAHEAD_SEARCH
-} from "graphql-module/queries";
-import {
-  ISearchAddressLinkNameResponse,
-  ISearchAddressLinkContainerResponse
-} from "types/ResponseTypes";
-import {
-  TypeAheadMessage,
-  TYPEAHEAD_REQUIRED_LENGTH,
-  FetchPolicy
-} from "constants/constants";
-import { getSelectOptionList, ISelectOption } from "utils";
+import { AddressLinksFilter } from "../AddressLinksFilter";
 
-interface IAddressLinksFilterProps {
+interface AddressLinksToolbar {
   filterValue: string;
   setFilterValue: (value: string) => void;
   filterNames: any[];
@@ -61,7 +28,7 @@ interface IAddressLinksFilterProps {
   addressName: string;
   namespace: string;
 }
-const AddressLinksToolbar: React.FunctionComponent<IAddressLinksFilterProps> = ({
+const AddressLinksToolbar: React.FunctionComponent<AddressLinksToolbar> = ({
   filterValue,
   setFilterValue,
   filterNames,
@@ -77,36 +44,6 @@ const AddressLinksToolbar: React.FunctionComponent<IAddressLinksFilterProps> = (
   namespace
 }) => {
   const { width } = useWindowDimensions();
-  const client = useApolloClient();
-  const [filterIsExpanded, setFilterIsExpanded] = React.useState<boolean>(
-    false
-  );
-  const [roleIsExpanded, setRoleIsExpanded] = React.useState<boolean>(false);
-  const [isSelectNameExpanded, setIsSelectNameExpanded] = React.useState<
-    boolean
-  >(false);
-  const [
-    isSelectContainerExpanded,
-    setIsSelectContainerExpanded
-  ] = React.useState<boolean>(false);
-  const [nameSelected, setNameSelected] = React.useState<string>();
-  const [containerSelected, setContainerSelected] = React.useState<string>();
-  const [nameOptions, setNameOptions] = React.useState<Array<ISelectOption>>();
-  const [containerOptions, setContainerOptions] = React.useState<
-    Array<ISelectOption>
-  >();
-  const [nameInput, setNameInput] = React.useState<string>("");
-  const [containerInput, setContainerInput] = React.useState<string>("");
-  const filterMenuItems = [
-    { key: "filterName", value: "Name" },
-    { key: "filterContainers", value: "Container" },
-    { key: "filterRole", value: "Role" }
-  ];
-
-  const roleMenuItems = [
-    { key: "roleSender", value: "Sender" },
-    { key: "roleReceiver", value: "Receiver" }
-  ];
 
   const sortMenuItems = [
     { key: "name", value: "Name", index: 2 },
@@ -114,418 +51,12 @@ const AddressLinksToolbar: React.FunctionComponent<IAddressLinksFilterProps> = (
     { key: "backlog", value: "Backlog", index: 4 }
   ];
 
-  const onAddInput = (event: any) => {
-    if (filterValue && filterValue === "Name") {
-      if (nameSelected && nameSelected.trim() !== "")
-        if (
-          filterNames.map(filter => filter.value).indexOf(nameSelected.trim()) <
-          0
-        ) {
-          setFilterNames([
-            ...filterNames,
-            { value: nameSelected.trim(), isExact: true }
-          ]);
-          setNameSelected(undefined);
-        }
-      if (!nameSelected && nameInput && nameInput.trim() !== "")
-        if (
-          filterNames.map(filter => filter.value).indexOf(nameInput.trim()) < 0
-        )
-          setFilterNames([
-            ...filterNames,
-            { value: nameInput.trim(), isExact: false }
-          ]);
-    } else if (filterValue && filterValue === "Container") {
-      if (containerSelected && containerSelected.trim() !== "")
-        if (
-          filterContainers
-            .map(filter => filter.value)
-            .indexOf(containerSelected.trim()) < 0
-        ) {
-          setFilterContainers([
-            ...filterContainers,
-            { value: containerSelected.trim(), isExact: true }
-          ]);
-          setContainerSelected(undefined);
-        }
-      if (!containerSelected && containerInput && containerInput.trim() !== "")
-        if (
-          filterContainers
-            .map(filter => filter.value)
-            .indexOf(containerInput.trim()) < 0
-        )
-          setFilterContainers([
-            ...filterContainers,
-            { value: containerInput.trim(), isExact: false }
-          ]);
-    }
-  };
-
-  const onFilterSelect = (event: any) => {
-    setFilterValue(event.target.value);
-    setFilterIsExpanded(!filterIsExpanded);
-  };
-  const onRoleSelect = (event: any) => {
-    setFilterRole(event.target.value);
-    setRoleIsExpanded(!roleIsExpanded);
-  };
-  const onNameSelectToggle = () => {
-    setIsSelectNameExpanded(!isSelectNameExpanded);
-  };
-  const onContainerSelectToggle = () => {
-    setIsSelectContainerExpanded(!isSelectContainerExpanded);
-  };
-  const onChangeNameData = async (value: string) => {
-    setNameOptions(undefined);
-    if (value.trim().length < TYPEAHEAD_REQUIRED_LENGTH) {
-      setNameOptions([]);
-      return;
-    }
-    const response = await client.query<ISearchAddressLinkNameResponse>({
-      query: RETURN_ALL_NAMES_OF_ADDRESS_LINK_FOR_TYPEAHEAD_SEARCH(
-        addressName,
-        namespace,
-        value.trim()
-      ),
-      fetchPolicy: FetchPolicy.NETWORK_ONLY
-    });
-    if (
-      response &&
-      response.data &&
-      response.data.addresses &&
-      response.data.addresses.addresses &&
-      response.data.addresses.addresses.length > 0 &&
-      response.data.addresses.addresses[0].links &&
-      response.data.addresses.addresses[0].links.links &&
-      response.data.addresses.addresses[0].links.links.length > 0
-    ) {
-      let obtainedList = [];
-      obtainedList = response.data.addresses.addresses[0].links.links.map(
-        (link: any) => {
-          return link.metadata.name;
-        }
-      );
-      //get list of unique records to display in the select dropdown based on total records and 100 fetched objects
-      const uniqueList = getSelectOptionList(
-        obtainedList,
-        response.data.addresses.addresses[0].links.total
-      );
-      if (uniqueList.length > 0) setNameOptions(uniqueList);
-    }
-  };
-  const onNameSelectFilterChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setNameInput(e.target.value);
-    onChangeNameData(e.target.value);
-    const options: React.ReactElement[] = nameOptions
-      ? nameOptions.map((option, index) => (
-          <SelectOption key={index} value={option} />
-        ))
-      : [];
-    return options;
-  };
-
-  const onChangeContainerData = async (value: string) => {
-    setContainerOptions(undefined);
-    if (value.trim().length < TYPEAHEAD_REQUIRED_LENGTH) {
-      setContainerOptions([]);
-      return;
-    }
-    const response = await client.query<ISearchAddressLinkContainerResponse>({
-      query: RETURN_ALL_CONTAINER_IDS_OF_ADDRESS_LINKS_FOR_TYPEAHEAD_SEARCH(
-        addressName,
-        namespace,
-        value.trim()
-      ),
-      fetchPolicy: FetchPolicy.NETWORK_ONLY
-    });
-    if (
-      response &&
-      response.data &&
-      response.data.addresses &&
-      response.data.addresses.addresses &&
-      response.data.addresses.addresses.length > 0 &&
-      response.data.addresses.addresses[0].links &&
-      response.data.addresses.addresses[0].links.links &&
-      response.data.addresses.addresses[0].links.links.length > 0
-    ) {
-      let obtainedList = [];
-      obtainedList = response.data.addresses.addresses[0].links.links.map(
-        (link: any) => {
-          return link.spec.connection.spec.containerId;
-        }
-      );
-      //get list of unique records to display in the select dropdown based on total records and 100 fetched objects
-      const uniqueList = getSelectOptionList(
-        obtainedList,
-        response.data.addresses.addresses[0].links.total
-      );
-      if (uniqueList.length > 0) setContainerOptions(uniqueList);
-    }
-  };
-
-  const onContainerSelectFilterChange = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    setContainerInput(e.target.value);
-    onChangeContainerData(e.target.value);
-    const options: React.ReactElement[] = containerOptions
-      ? containerOptions.map((option, index) => (
-          <SelectOption key={index} value={option} />
-        ))
-      : [];
-    return options;
-  };
-
-  const onNameSelect = (event: any, selection: string | SelectOptionObject) => {
-    setNameSelected(selection.toString());
-    setIsSelectNameExpanded(false);
-  };
-
-  const onContainerSelect = (
-    event: any,
-    selection: string | SelectOptionObject
-  ) => {
-    setContainerSelected(selection.toString());
-    setIsSelectContainerExpanded(false);
-  };
-
-  const onDelete = (
-    type: string | DataToolbarChip,
-    id: string | DataToolbarChip
-  ) => {
-    switch (type) {
-      case "Name":
-        if (filterNames && id) {
-          let index = filterNames
-            .map(filter => filter.value)
-            .indexOf(id.toString());
-          if (index >= 0) filterNames.splice(index, 1);
-          setFilterNames([...filterNames]);
-        }
-        break;
-      case "Container":
-        if (filterContainers && id) {
-          let index = filterContainers
-            .map(filter => filter.value)
-            .indexOf(id.toString());
-          if (index >= 0) filterContainers.splice(index, 1);
-          setFilterContainers([...filterContainers]);
-        }
-        break;
-      case "Role":
-        setFilterRole(undefined);
-        break;
-    }
-  };
-  const checkIsFilterApplied = () => {
-    if (
-      (filterNames && filterNames.length > 0) ||
-      (filterContainers && filterContainers.length > 0) ||
-      (filterRole && filterRole.trim() !== "")
-    ) {
-      return true;
-    }
-    return false;
-  };
   const onDeleteAll = () => {
     setFilterValue("Name");
     setFilterNames([]);
     setFilterContainers([]);
     setFilterRole(undefined);
   };
-
-  const toggleGroupItems = (
-    <>
-      <DataToolbarGroup variant="filter-group">
-        <DataToolbarFilter categoryName="Filter">
-          <Dropdown
-            id="ad-links-filter-dropdown"
-            position="left"
-            onSelect={onFilterSelect}
-            isOpen={filterIsExpanded}
-            toggle={
-              <DropdownToggle onToggle={setFilterIsExpanded}>
-                <FilterIcon />
-                &nbsp;
-                {filterValue}
-              </DropdownToggle>
-            }
-            dropdownItems={filterMenuItems.map(option => (
-              <DropdownItem
-                id={`ad-links-filter-dropdown-item${option.key}`}
-                key={option.key}
-                value={option.value}
-                itemID={option.key}
-                component={"button"}
-              >
-                {option.value}
-              </DropdownItem>
-            ))}
-          />
-        </DataToolbarFilter>
-        <>
-          <DataToolbarItem>
-            <DataToolbarFilter
-              chips={filterNames.map(filter => filter.value)}
-              deleteChip={onDelete}
-              categoryName="Name"
-            >
-              {filterValue && filterValue === "Name" && (
-                <InputGroup>
-                  <Select
-                    id="ad-links-filter-select-name"
-                    variant={SelectVariant.typeahead}
-                    aria-label="Select a Name"
-                    onToggle={onNameSelectToggle}
-                    onSelect={onNameSelect}
-                    onClear={() => {
-                      setNameSelected(undefined);
-                      setIsSelectNameExpanded(false);
-                    }}
-                    maxHeight="200px"
-                    selections={nameSelected}
-                    onFilter={onNameSelectFilterChange}
-                    isExpanded={isSelectNameExpanded}
-                    ariaLabelledBy={"typeahead-select-id"}
-                    placeholderText="Select name"
-                    isDisabled={false}
-                    isCreatable={false}
-                  >
-                    {nameOptions && nameOptions.length > 0 ? (
-                      nameOptions.map((option, index) => (
-                        <SelectOption
-                          key={index}
-                          value={option.value}
-                          isDisabled={option.isDisabled}
-                        />
-                      ))
-                    ) : nameInput.trim().length < TYPEAHEAD_REQUIRED_LENGTH ? (
-                      <SelectOption
-                        key={"invalid-input-length"}
-                        value={"Enter more characters"}
-                        disabled={true}
-                      />
-                    ) : (
-                      <SelectOption
-                        key={"no-results-found"}
-                        value={"No results found"}
-                        disabled={true}
-                      />
-                    )}
-                    {/* {} */}
-                  </Select>
-                  <Button
-                    id="ad-links-filter-search-name"
-                    variant={ButtonVariant.control}
-                    aria-label="search button for search name"
-                    onClick={onAddInput}
-                  >
-                    <SearchIcon />
-                  </Button>
-                </InputGroup>
-              )}
-            </DataToolbarFilter>
-          </DataToolbarItem>
-          <DataToolbarItem>
-            <DataToolbarFilter
-              chips={filterContainers.map(filter => filter.value)}
-              deleteChip={onDelete}
-              categoryName="Container"
-            >
-              {filterValue && filterValue === "Container" && (
-                <InputGroup>
-                  <Select
-                    id="ad-links-filter-select-container"
-                    variant={SelectVariant.typeahead}
-                    aria-label="Select a Container"
-                    onToggle={onContainerSelectToggle}
-                    onSelect={onContainerSelect}
-                    onClear={() => {
-                      setContainerSelected(undefined);
-                      setIsSelectContainerExpanded(false);
-                    }}
-                    maxHeight="200px"
-                    selections={containerSelected}
-                    onFilter={onContainerSelectFilterChange}
-                    isExpanded={isSelectContainerExpanded}
-                    ariaLabelledBy={"typeahead-select-id"}
-                    placeholderText="Select container"
-                    isDisabled={false}
-                    isCreatable={false}
-                  >
-                    {containerOptions && containerOptions.length > 0 ? (
-                      containerOptions.map((option, index) => (
-                        <SelectOption
-                          key={index}
-                          value={option.value}
-                          isDisabled={option.isDisabled}
-                        />
-                      ))
-                    ) : containerInput.trim().length <
-                      TYPEAHEAD_REQUIRED_LENGTH ? (
-                      <SelectOption
-                        key={"invalid-input-length"}
-                        value={TypeAheadMessage.MORE_CHAR_REQUIRED}
-                        disabled={true}
-                      />
-                    ) : (
-                      <SelectOption
-                        key={"no-results-found"}
-                        value={TypeAheadMessage.NO_RESULT_FOUND}
-                        disabled={true}
-                      />
-                    )}
-                    {/* {} */}
-                  </Select>
-                  <Button
-                    id="ad-links-filter-search-container"
-                    variant={ButtonVariant.control}
-                    aria-label="search button for search containers"
-                    onClick={onAddInput}
-                  >
-                    <SearchIcon />
-                  </Button>
-                </InputGroup>
-              )}
-            </DataToolbarFilter>
-          </DataToolbarItem>
-
-          <DataToolbarItem>
-            <DataToolbarFilter
-              chips={filterRole ? [filterRole] : []}
-              deleteChip={onDelete}
-              categoryName="Role"
-            >
-              {filterValue === "Role" && (
-                <Dropdown
-                  id="ad-links-filter-select-role"
-                  position="left"
-                  onSelect={onRoleSelect}
-                  isOpen={roleIsExpanded}
-                  toggle={
-                    <DropdownToggle onToggle={setRoleIsExpanded}>
-                      {filterRole || "Select Role"}
-                    </DropdownToggle>
-                  }
-                  dropdownItems={roleMenuItems.map(option => (
-                    <DropdownItem
-                      id={`ad-links-filter-select-option-role${option.key}`}
-                      key={option.key}
-                      value={option.value}
-                      itemID={option.key}
-                      component={"button"}
-                    >
-                      {option.value}
-                    </DropdownItem>
-                  ))}
-                />
-              )}
-            </DataToolbarFilter>
-          </DataToolbarItem>
-        </>
-      </DataToolbarGroup>
-    </>
-  );
 
   return (
     <DataToolbar
@@ -536,21 +67,19 @@ const AddressLinksToolbar: React.FunctionComponent<IAddressLinksFilterProps> = (
     >
       <DataToolbarContent>
         <>
-          <DataToolbarToggleGroup
-            toggleIcon={
-              <>
-                <FilterIcon />
-                {checkIsFilterApplied() && (
-                  <Badge key={1} isRead>
-                    {totalLinks}
-                  </Badge>
-                )}
-              </>
-            }
-            breakpoint="xl"
-          >
-            {toggleGroupItems}
-          </DataToolbarToggleGroup>
+          <AddressLinksFilter
+            filterValue={filterValue}
+            setFilterValue={setFilterValue}
+            filterNames={filterNames}
+            setFilterNames={setFilterNames}
+            filterContainers={filterContainers}
+            setFilterContainers={setFilterContainers}
+            filterRole={filterRole}
+            setFilterRole={setFilterRole}
+            totalLinks={totalLinks}
+            addressName={addressName}
+            namespace={namespace}
+          />
           {width < 769 && (
             <SortForMobileView
               sortMenu={sortMenuItems}
